+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added scaling funtions for different resolutions
 
+### Fixed
+
+- Fixed storybook's components
+
 ## [1.2.0] - 2023-12-06
 
 ### Added